cart api add

+++ src/main/java/com/example/Swigatto/model/Cart.java
@@ -22,7 +22,7 @@ public class Cart {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
 
-    int cartTotal;
+    double cartTotal;
 
     @OneToOne
     @JoinColumn
